Contour outlines from five leaves

A list of contour outlines from five leaves. It may be used as example in some functions such as efourier()

Format

A list with five objects

leaf_1

leaf_2

leaf_3

leaf_4

leaf_5

Each object is a data.frame with the coordinates for the outline perimeter

Source

Personal data. The images were obtained in the Flavia data set downlodable at https://flavia.sourceforge.net/

Author

Tiago Olivoto tiagoolivoto@gmail.com