Generate Custom Color Palette

This function generates a custom color palette using the specified colors and number of colors.

Usage

custom_palette(
  colors = c("#4B0055", "#00588B", "#009B95", "#53CC67", "yellow"),
  n = 100
)

Arguments

colors

A vector of colors to create the color palette. Default is c("steelblue", "salmon", "forestgreen").

n

The number of gradient colors in the color palette. Default is 100.

Value

A vector of colors representing the custom color palette.

Examples

# Generate a custom color palette with default colors and 10 colors
custom_palette()
#>   [1] "#4B0055" "#470357" "#440759" "#410A5B" "#3E0E5D" "#3B115F" "#381562"
#>   [8] "#351864" "#321C66" "#2F2068" "#2C236A" "#29276D" "#262A6F" "#232E71"
#>  [15] "#203173" "#1D3575" "#1A3877" "#173C7A" "#14407C" "#11437E" "#0E4780"
#>  [22] "#0B4A82" "#084E84" "#055187" "#025589" "#00588B" "#005B8B" "#005E8B"
#>  [29] "#00608C" "#00638C" "#00668D" "#00688D" "#006B8D" "#006E8E" "#00718E"
#>  [36] "#00738F" "#00768F" "#00798F" "#007B90" "#007E90" "#008191" "#008391"
#>  [43] "#008691" "#008992" "#008C92" "#008E93" "#009193" "#009493" "#009694"
#>  [50] "#009994" "#019B94" "#059D92" "#089F90" "#0BA18E" "#0FA38C" "#12A58A"
#>  [57] "#15A788" "#19A987" "#1CAB85" "#1FAD83" "#23AF81" "#26B17F" "#29B37D"
#>  [64] "#2DB57B" "#30B77A" "#33B978" "#37BB76" "#3ABD74" "#3EBF72" "#41C170"
#>  [71] "#44C36E" "#48C56D" "#4BC76B" "#4EC969" "#52CB67" "#58CD63" "#5FCF5F"
#>  [78] "#66D15B" "#6DD357" "#74D553" "#7AD74F" "#81D94A" "#88DB46" "#8FDE42"
#>  [85] "#96E03E" "#9DE23A" "#A4E436" "#ABE631" "#B2E82D" "#B9EA29" "#C0EC25"
#>  [92] "#C7EE21" "#CEF01D" "#D5F218" "#DCF414" "#E3F610" "#EAF80C" "#F1FA08"
#>  [99] "#F8FC04" "#FFFF00"

# Generate a custom color palette with specified colors and 20 colors
custom_palette(colors = c("blue", "red"), n = 20)
#>  [1] "#0000FF" "#0D00F1" "#1A00E4" "#2800D6" "#3500C9" "#4300BB" "#5000AE"
#>  [8] "#5D00A1" "#6B0093" "#780086" "#860078" "#93006B" "#A1005D" "#AE0050"
#> [15] "#BB0043" "#C90035" "#D60028" "#E4001A" "#F1000D" "#FF0000"

# example code
library(pliman)
custom_palette(n = 5)
#> [1] "#4B0055" "#00588B" "#009B95" "#53CC67" "#FFFF00"