List Computable Indexes Based on Available Bands

This function reads index equations from a CSV file included in the pliman package, determines which bands are used in each index equation, and checks which indexes can be computed based on the provided available bands.

Usage

pliman_indexes_ican_compute(available)

Arguments

available

A character vector of available bands (e.g., c("R", "G")).

Value

A data frame of indexes that can be computed with the available bands.

Examples

library(pliman)
available_bands <- c("R", "G")
computable_indexes <- pliman_indexes_ican_compute(available_bands)
print(computable_indexes)
#>    Index                  Equation  Band
#> 50   VIG               (G-R)/(G+R) MULTI
#> 65  ERVI              ((1.4*R)- G)   RGB
#> 66     G                         G   RGB
#> 70    GR                     G / R   RGB
#> 78 MGVRI       (G^2-R^2)/(G^2+R^2)   RGB
#> 82 NGRDI               (G-R)/(G+R)   RGB
#> 84     R                         R   RGB
#> 88  SAVI (1 + 0.5)*(G-R)/(G+R+0.5)   RGB
#> 89   SCI         (R - G) / (R + G)   RGB