Utilities for image resolution

Provides useful conversions between size (cm), number of pixels (px) and resolution (dpi)

dpi_to_cm(dpi)

cm_to_dpi(cm)

pixels_to_cm(px, dpi)

cm_to_pixels(cm, dpi)

Arguments

dpi
The image resolution in dots per inch.

cm
The size in centimeters.

px
The number of pixels.

Value

A numeric value or a vector of numeric values if the input data is a vector

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

library(pliman)
# Convert  dots per inch to centimeter
dpi_to_cm(c(1, 2, 3))

#> [1] 2.5400000 1.2700000 0.8466667

# Convert centimeters to dots per inch
cm_to_dpi(c(1, 2, 3))

#> [1] 0.3937008 0.7874016 1.1811024

# Convert centimeters to number of pixels with resolution of 96 dpi.
cm_to_pixels(c(1, 2, 3), 96)

#> [1]  37.79528  75.59055 113.38583

# Convert number of pixels to cm with resolution of 96 dpi.
pixels_to_cm(c(1, 2, 3), 96)

#> [1] 0.02645833 0.05291667 0.07937500