Import and export images

Import images from files and URLs and write images to files, possibly with batch processing.

image_import(
  image,
  ...,
  pattern = NULL,
  path = NULL,
  plot = FALSE,
  img_pattern = NULL
)

image_export(image, name, extension = NULL, subfolder = NULL, ...)

image_show(image)

image_pliman(image)

Arguments

image

For image_import(), a character vector of file names or URLs.

For image_export(), an Image object, an array or a list of images.

...
Alternative arguments passed to the corresponding functions from the jpeg, png, and tiff packages.

pattern
A pattern of file name used to identify images to be imported. For example, if pattern = "im" all images in the current working directory that the name matches the pattern (e.g., img1.-, image1.-, im2.-) will be imported as a list. Providing any number as pattern (e.g., pattern = "1") will select images that are named as 1.-, 2.-, and so on. An error will be returned if the pattern matches any file that is not supported (e.g., img1.pdf).

path
A character vector of full path names; the default corresponds to the working directory, getwd(). It will overwrite (if given) the path informed in image argument.

plot
Plots the image after importing? Defaults to FALSE.

img_pattern
Deprecated. Use pattern instead.

name
An string specifying the name of the image. It can be either a charactere with the image name (e.g., "img1") or name and extension (e.g., "img1.jpg"). If none file extension is provided, the image will be saved as a *.jpg file.

extension
When image is a list, extension can be used to define the extension of exported files. This will overwrite the file extensions given in image.

subfolder
Optional character string indicating a subfolder within the current working directory to save the image(s). If the folder doesn't exist, it will be created.

Value

image_import() returns a new Image object.

image_export() returns an invisible vector of file names.

image_pliman() returns a character string with the path to the example image installed with the package.

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

library(pliman)
(full_path <- image_pliman("sev_leaf.jpg"))

#> [1] "E:/Documents/R/win-library/4.1/pliman/tmp_images/sev_leaf.jpg"

(path <- file_dir(full_path))

#> [1] "E:/Documents/R/win-library/4.1/pliman/tmp_images"

(file <- basename(full_path))

#> [1] "sev_leaf.jpg"

image_import(image = full_path)

#> Image 
#>   colorMode    : Color 
#>   storage.mode : double 
#>   dim          : 490 557 3 
#>   frames.total : 3 
#>   frames.render: 1 
#> 
#> imageData(object)[1:5,1:6,1]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] 0.007843137 0.007843137 0.007843137 0.007843137 0.007843137 0.007843137
#> [2,] 0.007843137 0.007843137 0.007843137 0.007843137 0.007843137 0.007843137
#> [3,] 0.007843137 0.007843137 0.007843137 0.007843137 0.007843137 0.007843137
#> [4,] 0.007843137 0.007843137 0.007843137 0.007843137 0.007843137 0.007843137
#> [5,] 0.007843137 0.007843137 0.007843137 0.007843137 0.007843137 0.003921569

image_import(image = file, path = path)

#> Image 
#>   colorMode    : Color 
#>   storage.mode : double 
#>   dim          : 490 557 3 
#>   frames.total : 3 
#>   frames.render: 1 
#> 
#> imageData(object)[1:5,1:6,1]
#>             [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] 0.007843137 0.007843137 0.007843137 0.007843137 0.007843137 0.007843137
#> [2,] 0.007843137 0.007843137 0.007843137 0.007843137 0.007843137 0.007843137
#> [3,] 0.007843137 0.007843137 0.007843137 0.007843137 0.007843137 0.007843137
#> [4,] 0.007843137 0.007843137 0.007843137 0.007843137 0.007843137 0.007843137
#> [5,] 0.007843137 0.007843137 0.007843137 0.007843137 0.007843137 0.003921569