These functions applies common statistics to a list of objects, returning a numeric vector.

These functions applies common statistics to a list of objects, returning a numeric vector.

Usage

mean_list(x, ...)

sd_list(x, ...)

max_list(x, ...)

min_list(x, ...)

Arguments

x

A data.frame or matrix with numeric values.

...

Further arguments passed on to the R base function (e.g, mean(), sd(), etc.)

Value

A numeric vector.

Examples

mean_list(list(a = 1:10, b = 2:20))
#>    a    b 
#>  5.5 11.0