Sample images

Sample images installed with the pliman package

Format

*.jpg format

flax_leaves.jpg Flax leaves in a white background

flax_grains.jpg Flax grains with background light.

la_back.jpg A cyan palette representing the background of images la_pattern, la_leaves, and soybean_touch.

la_leaf.jpg A sample of the leaves in la_leaves

la_leaves.jpg Tree leaves with a sample of known area.

mult_leaves.jpg Three soybean leaflets with soybean rust symptoms.

objects_300dpi.jpg An image with 300 dpi resolution.

potato_leaves.jpg Three potato leaves, which were gathered from Gupta et al. (2020).

sev_leaf.jpg A soybean leaf with a blue background.

sev_leaf_nb.jpg A soybean leaf without background.

sev_back.jpg A blue palette representing the background of sev_leaf.

sev_healthy.jpg Healthy area of sev_leaf.

sev_sympt.jpg The symptomatic area sev_leaf.

shadow.jpg A shaded leaf, useful to test adaptive thresholding

soy_green.jpg Soybean grains with a white background.

soybean_grain.jpg A sample palette of the grains in soy_green.

soybean_touch.jpg Soybean grains with a cyan background touching one each other.

field_mosaic.jpg An UVA image from a soybean field.

Source

Personal data, Gupta et al. (2020).

References

Gupta, S., Rosenthal, D. M., Stinchcombe, J. R., & Baucom, R. S. (2020). The remarkable morphological diversity of leaf shape in sweet potato (Ipomoea batatas): the influence of genetics, environment, and G×E. New Phytologist, 225(5), 2183–2195. doi:10.1111/NPH.16286

Author

Tiago Olivoto tiagoolivoto@gmail.com