Import, display and export images

Import images from files and URLs, write images to files, and show images. *

image_import(image, ..., img_pattern = NULL)

image_export(image, name, ...)

image_show(image)

image_pliman(image)

Arguments

image

For image_import(), a character vector of file names or URLs.

For image_export(), an Image object, an array or a list of images.

...
Alternative arguments passed to the corresponding functions from the jpeg, png, and tiff packages.

img_pattern
A pattern of file name used to identify images to be imported. For example, if img_pattern = "im" all images in the current working directory that the name matches the pattern (e.g., img1.-, image1.-, im2.-) will be imported as a list. Providing any number as pattern (e.g., img_pattern = "1") will select images that are named as 1.-, 2.-, and so on.

name
An string specifying the name of the image.

Value

image_import() returns a new Image object.

image_export() returns an invisible vector of file names.

image_pliman() returns a character string with the path to the example image installed with the package.

Author

Tiago Olivoto tiagoolivoto@gmail.com

Examples

library(pliman)
img <- image_import(image_pliman("sev_leaf.jpg"))